pyproject.toml
[tool.poetry]
name = "sky-calendar"
version = "0.1.0"
description = ""
authors = ["Manuilov Daniil <manuilovdaniil@yandex.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "sky_calendar"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.1"
django-environ = "^0.10.0"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pre-commit = "^3.3.2"
ansible-vault-win = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
py_version = 311
profile = "django"
known_first_party = ["todolist", "core", "goals"]
skip = ["*/migrations/*"]
